+++ guillotina_elasticsearch/schema.py
@@ -100,4 +100,7 @@ Registered mapping: {field_mapping}
 
             schema_field_mappings[index_name] = catalog_info["__schema__"]
             mappings[index_name] = field_mapping
-    return {"properties": mappings}
+    return {
+        "properties": mappings,
+        "dynamic": app_settings.get("elasticsearch", {}).get("dynamic_mapping", False),
+    }