+++ CHANGELOG.rst
@@ -1,6 +1,9 @@
 6.0.3 (unreleased)
 ------------------
 
+- Disable dynamic mapping (as it was in v3.x)
+  [masipcat]
+
 - Fix tm is None when migration is cancelled
   [masipcat]